AG-5382-drag-drop-group-edit
