+++ packages/ag-grid-community/src/interfaces/iGroupEditService.ts
@@ -2,8 +2,8 @@ import type { RowsDrop } from '../dragAndDrop/rowDragTypes';
 import type { IRowNode } from './iRowNode';
 
 export interface IGroupEditService {
-    wouldCycle(row: IRowNode, newParent: IRowNode | null | undefined): boolean;
     canSetParent(rowsDrop: RowsDrop): boolean;
     isGroupingDrop(rowsDrop: RowsDrop): boolean;
     groupingEditDrop(rowsDrop: RowsDrop): boolean;
+    canDropRow(row: IRowNode, rowsDrop: RowsDrop): boolean;
 }
